requirements.txt
pandas==2.2.2
SQLAlchemy==2.0.32
pymysql==1.1.1
python-dotenv==1.0.1
